fix: move output generator to util
